Testing Mode + Direction Fixes
Added an aditional TestMotors class that is a copy of our main class
that has each individual motor mapped to the d-pad. This was then used
to determine which motor is which, so that the joystick could be
propperly mapped. The robot now moves propperly in the direction that
the joystick is aimed at.